Don't let user exist in multiple tenants when using custom sso
